tests/wpt/tests/css/css-view-transitions/navigation/at-rule-multiple-rules.html

<!DOCTYPE html>
<title>View Transitions: Multiple @view-transition, last one wins.</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-2/">
<link rel="author" href="mailto:bokan@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@view-transition {
  navigation: auto;
}
@view-transition {
  navigation: none;
}
</style>
<script>
const params = new URLSearchParams(location.search);
const is_new_page = params.has('new');

if (!is_new_page) {
  onload = () => requestAnimationFrame(() => requestAnimationFrame(() => {
      location.replace(location.href + '?new');
  }));
} else {
  promise_test(() => {
    return new Promise((resolve) => {
        addEventListener('pagereveal', resolve);
    }).then(event => {
        assert_equals(event.viewTransition, null,
            "ViewTransition must not be triggered.");
    });
  });
}
</script>
